Modify function to accept props
